NAV-8-moved auth service to auth folder
